+++ Src/Private/Get-AbrVbrBackupServerInfo.ps1
@@ -6,7 +6,7 @@ function Get-AbrVbrBackupServerInfo {
     .DESCRIPTION
         Documents the configuration of Veeam VBR in Word/HTML/Text formats using PScribo.
     .NOTES
-        Version:        0.7.1
+        Version:        0.8.0
         Author:         Jonathan Colon
         Twitter:        @jcolonfzenpr
         Github:         rebelinux
@@ -519,49 +519,6 @@ function Get-AbrVbrBackupServerInfo {
                         }
                         catch {
                             Write-PscriboMessage -IsWarning "Backup Server Service Status Section: $($_.Exception.Message)"
-                        }
-                        try {
-                            Write-PScriboMessage "Infrastructure Backup Server InfoLevel set at $($InfoLevel.Infrastructure.BackupServer)."
-                            if ($InfoLevel.Infrastructure.BackupServer -ge 3) {
-                                $NetStats = Get-VeeamNetStat -Session $PssSession | Where-Object { $_.ProcessName -Like "*veeam*" } | Sort-Object -Property State,LocalPort
-                                Write-PscriboMessage "Collecting Backup Server Network Statistics from $($BackupServer.Name)."
-                                if ($NetStats) {
-                                    Section -Style Heading4 "HealthCheck - Network Statistics" {
-                                        $OutObj = @()
-                                        foreach ($NetStat in $NetStats) {
-                                            try {
-                                                $inObj = [ordered] @{
-                                                    'Proto' = $NetStat.Protocol
-                                                    'Local IP' = $NetStat.LocalAddress
-                                                    'Local Port' = $NetStat.LocalPort
-                                                    'Remote IP' = $NetStat.RemoteAddress
-                                                    'Remote Port' = $NetStat.RemotePort
-                                                    'State' = $NetStat.State
-                                                    'Process Name' = $NetStat.ProcessName
-                                                    'PID' = $NetStat.PID
-                                                }
-                                                $OutObj += [pscustomobject]$inobj
-                                            }
-                                            catch {
-                                                Write-PscriboMessage -IsWarning "Backup Server Network Statistics $($NetStat.Protocol) Section: $($_.Exception.Message)"
-                                            }
-                                        }
-
-                                        $TableParams = @{
-                                            Name = "HealthCheck - Network Statistics - $($BackupServer.Name.Split(".")[0])"
-                                            List = $false
-                                            ColumnWidths = 8, 16, 8, 16, 9, 16, 19, 8
-                                        }
-                                        if ($Report.ShowTableCaptions) {
-                                            $TableParams['Caption'] = "- $($TableParams.Name)"
-                                        }
-                                        $OutObj | Table @TableParams
-                                    }
-                                }
-                            }
-                        }
-                        catch {
-                            Write-PscriboMessage -IsWarning "Backup Server Network Statistics Section: $($_.Exception.Message)"
                         }
                     }
                 }