+++ Src/Private/Get-AbrVbrSecInfraHard.ps1
@@ -6,7 +6,7 @@ function Get-AbrVbrSecInfraHard {
     .DESCRIPTION
         Documents the configuration of Veeam VBR in Word/HTML/Text formats using PScribo.
     .NOTES
-        Version:        0.6.0
+        Version:        0.8.0
         Author:         Jonathan Colon
         Twitter:        @jcolonfzenpr
         Github:         rebelinux